# apps/web/openapi.yaml
openapi: 3.1.0
info:
  title: Mintenance Platform API
  description: |
    RESTful API for the Mintenance contractor discovery and job marketplace platform.
    
    ## Authentication
    Most endpoints require authentication via JWT Bearer token. Obtain a token by calling `/api/auth/login`.
    
    ## Rate Limiting
    API endpoints are rate limited. AI/ML endpoints have stricter limits (5 requests/minute).
    Exceeding rate limits returns a 429 status code.
    
    ## Security
    - All endpoints use HTTPS
    - CSRF protection on mutation endpoints
    - Rate limiting on all public endpoints
  version: 1.2.4
  contact:
    name: Mintenance API Support
    email: api-support@mintenance.com
  license:
    name: Proprietary
    url: https://mintenance.com/terms
servers:
  - url: https://api.mintenance.com
    description: Production server
  - url: https://staging-api.mintenance.com
    description: Staging server
  - url: http://localhost:3000
    description: Development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/login
    
    CSRFToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: CSRF token for mutation endpoints
  schemas:
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
          additionalProperties: true
...
    AIAssessment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        job_id:
          type: string
          format: uuid
        damage_type:
          type: string
        severity:
          type: string
          enum: [minor, moderate, severe, critical]
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
        estimated_cost_min:
          type: number
          format: float
        estimated_cost_max:
          type: number
          format: float
        recommendations:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
...
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: UNAUTHORIZED
            message: Authentication required
...
    RateLimitError:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: RATE_LIMIT_EXCEEDED
            message: Too many requests. Please try again later.
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
...
paths:
...
  /api/building-surveyor/assess:
    post:
      tags:
        - AI
      summary: Building damage assessment
      description: AI-powered building damage assessment using GPT-4 Vision
      operationId: buildingAssessment
      security:
        - BearerAuth: []
        - CSRFToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - photos
              properties:
                photos:
                  type: array
                  items:
                    type: string
                    format: uri
                  minItems: 1
                  maxItems: 10
                context:
                  type: object
                  properties:
                    property_type:
                      type: string
                    building_age:
                      type: integer
                    previous_damage:
                      type: boolean
      responses:
        '200':
          description: Assessment complete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIAssessment'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'

  # =====================
  # MESSAGES
  # =====================
...